### 2022-06-29 #4526 @editorialbot
[PRE REVIEW]: Accelerating Parallel Operation for Compacting Selected Elements on GPUs
**Submitting author:** @yogi-tud (<a href="http://orcid.org/0000-0001-7898-0502">Johannes Fett</a>)
**Repository:** https://github.com/yogi-tud/space_gpu
**Branch with paper.md** (empty if default branch):
**Version:** 1.0
