+++ paper.bib
@@ -11,7 +11,8 @@
   number={2},
   pages={208--226},
   year={2017},
-  publisher={IJNC Editorial Committee}
+  publisher={IJNC Editorial Committee},
+  doi={http://doi.org/10.15803/ijnc.7.2_208}
 }
 
 @misc{space,
