### 2022-07-05 #4526 @yogi-tud
(seems like check references did not work before) I updated the bib files.
