### 2022-07-05 #4526 @yogi-tud
Ok is seems i should not include http:// in the DOI as the bot adds https itself. ill fix that right away.
