+++ paper.bib
@@ -12,7 +12,7 @@
   pages={208--226},
   year={2017},
   publisher={IJNC Editorial Committee},
-  doi={doi.org/10.15803/ijnc.7.2_208}
+  doi={10.15803/ijnc.7.2_208}
 }
 
 @misc{space,
