### 2022-07-19 yogi-tud/space_gpu#1 @robertszafa
[Documentation] Add short example usage 
A short description of how to use the gpu_compressstore2 binary and an example run would be helpful.

Checkpoint:
Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).
openjournals/joss-reviews#4589
