### 2022-07-19 yogi-tud/space_gpu#2 @robertszafa
[Documentation] Add README
A README (such that it can be displayed on the github repo) would be helpful for new users.
Either putting the contents of overview.pdf into a README format or linking to overview.pdf from the README should do the trick.

Checkpoint:
Functionality documentation: Is the core functionality of the software documented to a satisfactory level
https://github.com/openjournals/joss-reviews/issues/4589
