### 2022-07-19 yogi-tud/space_gpu#3 @robertszafa
[Documentation] Add contribution guidelines.
A short CONTRIBUTE file in the root folder should be enough.

Checkpoint:

openjournals/joss-reviews#4589
