### 2022-07-19 yogi-tud/space_gpu#3 @yogi-tud
I added a contribution file which focuses on the idea of adding another gpu compaction algorithm and comparing it to SPACE and Nvidia CUB.
