### 2022-07-19 yogi-tud/space_gpu#1 @yogi-tud
A short example and overview of all runtime parameters have been added to readme
