### 2022-07-19 #4589 @yogi-tud
Hi @robertszafa
Thanks for the input. I have added the documentation and a readme file providing an example.
Also some information about contributing in a file called CONTRIBUTE.
