### 2022-07-20 yogi-tud/space_gpu#3 @wimvanderbauwhede
Could you make this a little bit more complete please? The JOSS guidance for "Community guidelines" is:

There should be clear guidelines for third-parties wishing to:

    Contribute to the software
    Report issues or problems with the software
    Seek support
