### 2022-07-20 yogi-tud/space_gpu#2 @wimvanderbauwhede
Could you please either link to "overview.pdf" or (better) integrate it into the README?
