### 2022-07-20 yogi-tud/space_gpu#2 @yogi-tud
I linked to the overview pdf.
I think it is better to link it instead of having a multi page markdown readme.
