### 2022-07-20 yogi-tud/space_gpu#3 @yogi-tud
I have added sections for reporting issues and seeking support at:
https://github.com/yogi-tud/space_gpu/blob/main/CONTRIBUTE.md
