### 2022-07-21 yogi-tud/space_gpu#2 @robertszafa
I'm okay with this. Unless @wimvanderbauwhede has any comments, I will close this issue.
