### 2022-07-21 yogi-tud/space_gpu#2 @wimvanderbauwhede
OK for me!
