+++ paper.bib
@@ -17,12 +17,13 @@
 
 @misc{space,
     title={Accelerating Parallel Operation for Compacting Selected Elements on GPUs},
-     author = {Johannes Fett, Urs Kober, Christian Schwarz, Dirk Habich, Wolfgang Lehner},
+    author = {Johannes Fett and Urs Kober and Christian Schwarz and Dirk Habich and Wolfgang Lehner},
+    year = 2022,
     howpublished = {\url{https://github.com/yogi-tud/space_gpu/blob/main/Euro_Pars_2022_Compaction_CRC.pdf}}
 }
 
 @misc{overview,
     title={Overview},
-    author = {Johannes Fett, Urs Kober, Christian Schwarz, Dirk Habich, Wolfgang Lehner},
+    author = {Johannes Fett and Urs Kober and Christian Schwarz and Dirk Habich and Wolfgang Lehner},
     howpublished = {\url{https://github.com/yogi-tud/space_gpu/blob/main/overview.pdf}}
 }
