### 2022-07-22 #4589 @danielskatz
@yogi-tud - I'm suggesting a number of small changes to your paper in https://github.com/yogi-tud/space_gpu/pull/4.  In addition, I think a reference would be useful for "the NVIDIA-supplied CUB library"  Once this is done, we can move to the final acceptance steps
