+++ paper.bib
@@ -1,5 +1,5 @@
 @misc{cub,
-    author = {{CUB: Main Page}},
+    author = {{NVIDIA CUB library}},
     howpublished = {\url{https://nvlabs.github.io/cub/index.html}}
 }
 
