### 2022-07-22 #4589 @yogi-tud
@danielskatz There is already a reference to the cub main page. I renamed it to "NVIDIA CUB library" in the bib.
Is that okay?
