### 2022-07-22 #4589 @yogi-tud
I have a question about meta data in Zenodo.
Should it be "publication" or "software"?
