### 2022-07-22 #4589 @yogi-tud
https://doi.org/10.5281/zenodo.6884000

I uploaded the 1.0 release which was just created from the github repo to zenodo.
As my funding institutions were not listes as possible options, I did not add them unter the optional funding category.
