### 2022-07-22 #4589 @danielskatz
Also @wimvanderbauwhede, do we know the DOI for the Euro-Par paper?  If so, we could also mention that in the left column, as we do with AAS-partnered papers, such as in https://joss.theoj.org/papers/10.21105/joss.04417
