### 2022-07-25 #4589 @wimvanderbauwhede
@danielskatz I have the DOIs for the conference papers but they are not live yet so if we put them in the JOSS paper they won't work yet.
