+++ paper.md
@@ -23,6 +23,7 @@ affiliations:
    index: 1
 date: 28 June 2022
 bibliography: paper.bib
+europar-doi: 10.1007/978-3-031-12597-3_12
 ---
 
 # Summary
