### 2022-08-03 #4589 @arfon
@yogi-tud – could you please merge this PR to associate this submission with your Europar submission? https://github.com/yogi-tud/space_gpu/pull/5
