### 2022-08-03 #4589 @arfon
What do you think @wimvanderbauwhede and @danielskatz?: https://github.com/openjournals/joss-papers/blob/e881a2735349adace33e66f83a60aab1852227a7/joss.04589/10.21105.joss.04589.pdf
