### 2022-08-03 #4589 @danielskatz
looks good to me - https://joss.theoj.org/papers/10.21105/joss.04589
